first version of full process
